third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/open-close/open-features-non-integer-top-expected.txt

This is a testharness.js-based test.
PASS features "top=/104" should NOT set "top=104"
PASS features "top=_104" should NOT set "top=104"
PASS features "top=L104" should NOT set "top=104"
FAIL features "top=105.5" should set "top=105" assert_equals: "top=105.5 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "top=105.32" should set "top=105" assert_equals: "top=105.32 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "top=105LLl" should set "top=105" assert_equals: "top=105LLl value after first non-digit will be ignored" expected 105 but got 0
FAIL features "top=105^4" should set "top=105" assert_equals: "top=105^4 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "top=105*3" should set "top=105" assert_equals: "top=105*3 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "top=105/5" should set "top=105" assert_equals: "top=105/5 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "top=105  " should set "top=105" assert_equals: "top=105   value after first non-digit will be ignored" expected 105 but got 0
FAIL features "top=105e1" should set "top=105" assert_equals: "top=105e1 value after first non-digit will be ignored" expected 105 but got 0
FAIL features "top=105e-1" should set "top=105" assert_equals: "top=105e-1 value after first non-digit will be ignored" expected 105 but got 0
Harness: the test ran to completion.
